==> 03_stan_results.Rmd <==
---
title: "Salamander Analysis- Stan Results"
author: "Will Townes"
date: "11/19/2017"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ggplot2)
library(rstan)
library(cowplot)
load("data/bl_stanfit.rda") #blfit, bl
load("data/po_stanfit.rda") #pofit, po
dat<-read.csv("data/merged_data.csv")
dat$saltype<-plyr::mapvalues(dat$saltype, from=c("blnh","ponh","rbnh"), to=c("BL","PO","RB"))

#check for stan errors
if(blfit@mode != 0) stop("bl_stanfit did not converge, try again!")
if(pofit@mode != 0) stop("po_stanfit did not converge, try again!")
```

## Negative Binomial Model

### Posterior Mean of Temperature Functions

Big levels location

```{r}
tof<-10 #temperature offset (Celsius)
bl_mn<-summary(blfit,pars="beta",probs=NULL)$summary[,"mean"]
names(bl_mn)<-colnames(bl$X)
bl_bl<-function(t){
  res<-bl_mn["(Intercept)"]+(t-tof)*bl_mn["temp"]+(t-tof)^2*bl_mn["temp2"]
  exp(res)
}
bl_rb<-function(t){
  res<-bl_mn["(Intercept)"]+bl_mn["saltyperbnh"]+
    (t-tof)*(bl_mn["temp"]+bl_mn["saltyperbnh:temp"])+
    (t-tof)^2*(bl_mn["temp2"]+bl_mn["saltyperbnh:temp2"])
  exp(res)
}
bl_col<-c(BL="#009E73",RB="#D55E00")
bl_plt<-ggplot(subset(dat,site=="BL"),aes(x=temp,colour=saltype))+ylab("count of adult salamanders")+scale_color_manual(values=bl_col)+xlim(c(0,25))+geom_point(aes(y=salcount))+stat_function(fun=bl_rb,colour=bl_col["RB"])+stat_function(fun=bl_bl,colour=bl_col["BL"])+guides(colour=guide_legend(title="species"))+xlab("temperature (C)")+coord_trans(y="log1p")
show(bl_plt+ggtitle("Big Levels vs. Redbacks"))
ggsave(file="plots/negbinom_bl_vs_rb.png",width=6,height=4)
```

peaks of otter location

```{r}
po_mn<-summary(pofit,pars="beta",probs=NULL)$summary[,"mean"]
names(po_mn)<-colnames(po$X)
po_po<-function(t){
  res<-po_mn["(Intercept)"]+(t-tof)*po_mn["temp"]+(t-tof)^2*po_mn["temp2"]
  exp(res)
}
po_rb<-function(t){
  res<-po_mn["(Intercept)"]+po_mn["saltyperbnh"]+
    (t-tof)*(po_mn["temp"]+po_mn["saltyperbnh:temp"])+
    (t-tof)^2*(po_mn["temp2"]+po_mn["saltyperbnh:temp2"])
  exp(res)
}
po_col<-c(PO="#0072B2",RB="#D55E00")
po_plt<-ggplot(subset(dat,site=="PO"),aes(x=temp,colour=saltype))+ylab("count of adult salamanders")+scale_color_manual(values=po_col)+xlim(c(0,25))+geom_point(aes(y=salcount))+stat_function(fun=po_rb,colour=po_col["RB"])+stat_function(fun=po_po,colour=po_col["PO"])+guides(colour=guide_legend(title="species"))+xlab("temperature (C)")+coord_trans(y="log1p")
show(po_plt+ggtitle("Peaks of Otter vs. Redbacks"))
ggsave(file="plots/negbinom_po_vs_rb.png",width=6,height=4)
```

Combining the two locations into a single plot

```{r}
plot_grid(bl_plt,po_plt,nrow=2)
```

As shown above, the negative binomial model suggests there is a small but significant difference in the temperature profiles for the red backed and peaks of otter salamanders. In particular, the peaks of otter salamanders seem to prefer slightly higher temperatures than the red backs. There does not appear to be a significant difference in the temperature profiles for the big levels versus the red backed salamanders.

### Alternative Parameterization

Dr. Marsh's model is `SALCOUNT (neg.bin) ~ SPECIES + temp + temp^2 + SPECIES*temp + SPECIES*temp^2 +soilMS + (~1|date)`. Consider the quadratic function on the log scale, with $\mu$ the mean abundance and $x$ representing temperature:

$$\log\mu = ax^2+bx+c$$

There are three parameters, $a,b,c$. Using some algebra, we can re-parameterize this relationship as

$$\log\mu = -s\left(\frac{x-q}{w/2}\right)^2+s$$

The parameters $q,r,s$ are more interpretable: $q=-b/(2a)$ is the temperature at which abundance reaches its peak. The maximum abundance is given by $e^s$ where $s=c-b^2/(4a)$ and the range of temperatures where log-abundance is expected to be more than zero has width $w=\sqrt{b^2-4ac}/|a|$ (ie mean abundance expected to be more than one). We can also obtain the range of temperatures where mean abundance is greater than $k$ by $\sqrt{b^2-4a(c-\log k)}/|a|$. Based on this transformation, we can compare the abundance curves for different species by looking at contrasts in these parameters. For example, we can ask

* Do the two species reach maximum abundance at different temperatures? (comparing $q$)
* Does one species have a wider temperature range? (comparing $w$)
* Is one species more abundant than the other (comparing $s$)

To improve numerical stability while working with the quadratic temperature terms, we would prefer to work with the offset temperature $x-k$ where $k$ is some baseline temperature. Based on the above plots, we could set $k=10$ (ie temperature relative to 10 degrees celsius). Then the temperature at which abundance reaches its peak changes to $q+k$ instead of $q$. The other two parameters retain the same interpretation when calculated on the offset scale.

In a previous step, we already transformed temperature by subtracting it from the offset of 10 degrees Celsius for both the BL and PO datasets. We also transformed soil moisture by subtracting off the median at each site. So, if we set soil moisture to zero it is the same as interpreting the parameters at the median level for each site.

```{r}
transform_temp_pars<-function(a,b,c){
  q<- -b/(2*a)
  w<- sqrt(b^2 - 4*a*c)/abs(a)
  s<- c - b^2/(4*a)
  data.frame(q=q, s=s, w=w)
}
extract_transform<-function(stfit,stdat,tof=10,qlo=0,qhi=25,a_hi=-.001){
  #stfit a stan fit object
  #stdat the data used by stan
  #tof: offset for temperature, affects interpretation of "q" param
  #qlo and qhi are bounds on what a reasonable "peak abundance" temperature could be (Celsius)
  #returns: a data frame of samples from posterior
  #excludes samples with "unreasonable" q values and NaN values for w
  #excludes samples for which the quadratic coefficient was above "a_hi"
  #this can enforce negative concavity of the samples
  #those samples had numerical problems leading to ill-conditioned values
  #also computes q_dif and w_dif which are differences between red-backed and endemic parameters
  #extract sample draws for fixed effects
  res<-as.data.frame(stfit,pars="beta")
  pn<-colnames(res)<-colnames(stdat$X) #parameter names
  bad_a<-res$temp2 > a_hi | res$temp2+res$`saltyperbnh:temp2` > a_hi
  res<-res[!bad_a,]
  en<-transform_temp_pars(res$temp2,res$temp,res$`(Intercept)`)
  en$q<-en$q+tof
  colnames(en)<-paste0(colnames(en),"_en")
  #had warnings about NAN in temp_range due to sqrt of negative number
  #apply(saltype_base,2,function(x){sum(is.na(x))})
  #hist(res$temp2) #too close to zero!
  res<-cbind(res,en)
  rb<-transform_temp_pars(res$temp2+res$`saltyperbnh:temp2`,res$temp+res$`saltyperbnh:temp`,res$`(Intercept)`+res$saltyperbnh)
  rb$q<-rb$q+tof
  #apply(saltype_other,2,function(x){sum(is.na(x))})
  #hist(res$temp2+res$`saltyperbnh:temp2`) #too close to zero!
  colnames(rb)<-paste0(colnames(rb),"_rb")
  res<-cbind(res,rb)
  #exclude samples where temperature range was NaN (not enough curvature to define a maximum in these samples).
  #also exclude samples where q was less than 0 or more than 25, these are ill-defined values resulting from the sampled quadratic curves being too flat
  #can think of this maybe as having a "prior" on q that it must be between 0 and 25
  find_bad_q<-function(q,lo=qlo,hi=qhi){q<lo | q>hi}
  bad_samples<-with(res,is.nan(w_en) | is.nan(w_rb) | find_bad_q(q_en) | find_bad_q(q_rb))
  res<-res[!bad_samples,]
  res$q_dif<-res$q_rb - res$q_en
  res$w_dif<-res$w_rb - res$w_en
  res
}
```

### Big Levels Results

Examining the 90% and 95% credible intervals, there does not appear to be a significant difference between red-backed and big levels salamanders in terms of peak temperature preference (q). However, it appears that at this site, the red-backed salamanders have a significantly narrower temperature range compared to the endemic big levels species.
```{r}
resbl<-extract_transform(blfit,bl)
hist(resbl$q_dif)
quantile(resbl$q_dif,c(.025,.975))
quantile(resbl$q_dif,c(.05,.95))
mean(resbl$q_dif) # less than zero means rbnh likes lower temps
mean(resbl$q_dif<0) #posterior prob. that rbnh likes lower temps than blnh
hist(resbl$w_dif)
quantile(resbl$w_dif,c(.025,.975))
```

### Peaks of Otter Results

Examining the 90% and 95% credible intervals, there does not appear to be a significant difference between red-backed and peaks of otter salamanders in terms of peak temperature preference (q). However, it appears that at this site, the red-backed salamanders have a significantly wider temperature range compared to the endemic peaks of otter species.
```{r}
respo<-extract_transform(pofit,po)
hist(respo$q_dif)
quantile(respo$q_dif,c(.025,.975))
quantile(respo$q_dif,c(.05,.95))
mean(respo$q_dif) #less than zero means rbnh likes lower temps
mean(respo$q_dif<0) #posterior prob. that rbnh likes lower temps than ponh
hist(respo$w_dif)
quantile(respo$w_dif,c(.025,.975))
```

old code, keeping for reference

```{r, eval=FALSE}
other_saltype_name<-pn[3]
#saltype_other<-as.data.frame(with(res,transform_temp_pars(

#res<-cbind(res,

res<-summary(stfit)$summary[,c("mean","2.5%","97.5%")]
beta<-as.data.frame(res[startsWith(rownames(res),"beta"),])
beta$signif<-beta[,"2.5%"]*beta[,"97.5%"]>0 #CI spans zero?
rownames(beta)<-colnames(stdat$X)
#beta is matrix with rows=parameters and cols=mean and conf. interval.
sigma<-res["sigma",]
rand_ints<-res[startsWith(rownames(res),"rand_ints"),]
hist(exp(rand_ints[,"mean"])) #visualize random intercept distribution
phi<-res[startsWith(rownames(res),"phi"),]
#visualize the implied curves
temp_curve<-function(x,species=c("blnh","ponh","rbnh"),temp_offset=10,logscale=TRUE){
  #prediction for log-mean function
  species<-match.arg(species)
  xt<-x-temp_offset
  b<-beta[,"mean"]
  names(b)<-rownames(beta)
  lmu<-b["(Intercept)"]+b["soilWC"]*wc_median+b["temp"]*xt+b["temp2"]*xt^2
  #if(species=="rbnh"){
  #  lmu<-lmu+b["saltyperbnh"]+b["saltyperbnh:temp"]*xt+b["saltyperbnh:temp2"]*xt^2
  #}
  if(species=="blnh"){
    lmu<-lmu+b["saltypeblnh"]+b["saltypeblnh:temp"]*xt+b["saltypeblnh:temp2"]*xt^2
  }
  if(!logscale) return(exp(lmu))
  lmu
}

ggplot(dat[dat$site=="BL",],aes(x=temp,y=salcount,colour=saltype))+geom_point() + stat_function(fun=function(x){temp_curve(x,"blnh",logscale=FALSE)},colour="black") + stat_function(fun=function(x){temp_curve(x,"rbnh",logscale=FALSE)},colour="black")+xlab("temperature(C)")+ylab("salamander abundance (count)")

knitr::kable(beta)
knitr::kable(phi)
```
